move code specific to nuxt webpack context into plugin.utils
don't use ufo alias in pure module context as that should be handled
natively by npm/node
