+++ src/templates/plugin.routing.js
@@ -1,7 +1,7 @@
 import './middleware'
 import Vue from 'vue'
 import { Constants, nuxtOptions, options } from './options'
-import { getDomainFromLocale } from './utils-common'
+import { getDomainFromLocale } from './plugin.utils'
 // @ts-ignore
 import { withoutTrailingSlash, withTrailingSlash } from '~i18n-ufo'
